so far services payment tests working
