+++ test/suites/common-tanssi/services-payment/test_service_payment_removes_tank_money_and_refunds.ts
@@ -18,9 +18,13 @@ describeSuite({
         let refundAddress;
         let balanceTankBefore;
         let purchasedCredits;
+        let registerAlias;
         beforeAll(async () => {
             polkadotJs = context.polkadotJs();
             alice = context.keyring.alice;
+            const runtimeName = polkadotJs.runtimeVersion.specName.toString();
+            registerAlias = runtimeName.includes("light") ? polkadotJs.tx.containerRegistrar : polkadotJs.tx.registrar;
+
             refundAddress = generateKeyringPair("sr25519");
             const tx2001OneSession = polkadotJs.tx.servicesPayment.setBlockProductionCredits(paraId2001, 0);
             await context.createBlock([await polkadotJs.tx.sudo.sudo(tx2001OneSession).signAsync(alice)]);
@@ -51,7 +55,7 @@ describeSuite({
             title: "On deregistration we refund the address",
             test: async function () {
                 // We deregister the chain
-                const deregister2001 = polkadotJs.tx.sudo.sudo(polkadotJs.tx.registrar.deregister(paraId2001));
+                const deregister2001 = polkadotJs.tx.sudo.sudo(registerAlias.deregister(paraId2001));
                 await context.createBlock([await deregister2001.signAsync(alice)]);
                 // Check that after 2 sessions, tank is empty and chain is deregistered
                 await jumpSessions(context, 2);
