+++ test/suites/dev-tanssi-relay/services-payment/test_services_payment_collator_tip.ts
@@ -3,7 +3,7 @@ import { describeSuite, expect, beforeAll } from "@moonwall/cli";
 import { ApiPromise } from "@polkadot/api";
 import { KeyringPair } from "@moonwall/util";
 import { fetchCollatorAssignmentTip, jumpSessions } from "util/block";
-import { paraIdTank } from "util/payment";
+
 describeSuite({
     id: "CT0608",
     title: "Services payment collator assignment tip test suite",
@@ -11,12 +11,10 @@ describeSuite({
     testCases: ({ it, context }) => {
         let polkadotJs: ApiPromise;
         let alice: KeyringPair;
-        let collatorAssignmentAlias;
+
         beforeAll(async () => {
             polkadotJs = context.polkadotJs();
             alice = context.keyring.alice;
-            const runtimeName = polkadotJs.runtimeVersion.specName.toString();
-            collatorAssignmentAlias = runtimeName.includes("light") ? polkadotJs.query.tanssiCollatorAssignment : polkadotJs.query.collatorAssignment;
         });
         it({
             id: "E01",
@@ -32,22 +30,17 @@ describeSuite({
                 await context.createBlock([await tx.signAsync(alice)]);
 
                 const txMaxTip = polkadotJs.tx.servicesPayment.setMaxTip(paraId, tip);
-                await context.createBlock([await polkadotJs.tx.sudo.sudo(txMaxTip).signAsync(alice)]);
+                // In genesis we have 4 collators, hence if we make 4 collators per para, we make sure the one
+                // with priority gets them
+                const changeCollatorsPerChain = polkadotJs.tx.collatorConfiguration.setCollatorsPerContainer(4);
+                await context.createBlock([await polkadotJs.tx.sudo.sudo(polkadotJs.tx.utility.batchAll([txMaxTip, changeCollatorsPerChain])).signAsync(alice)]);
                 await jumpSessions(context, 2);
 
-                const collators = await collatorAssignmentAlias.collatorContainerChain();
-
-                const balanceTank = (
-                    await polkadotJs.query.system.account(paraIdTank(paraId))
-                ).data.free.toBigInt();
-                
-                console.log(balanceTank)
-                console.log(paraIdTank(paraId))
-                
+                const collators = await polkadotJs.query.tanssiCollatorAssignment.collatorContainerChain();
                 expect(
                     collators.toJSON().containerChains[paraId].length,
-                    `Container chain ${paraId} should have 2 collators`
-                ).toBe(2);
+                    `Container chain ${paraId} should have 4 collators`
+                ).toBe(4);
 
                 const events = await polkadotJs.query.system.events();
                 const tipEvent = fetchCollatorAssignmentTip(events);
