+++ test/util/block.ts
@@ -4,6 +4,8 @@ import { filterAndApply } from "@moonwall/util";
 import { ApiPromise } from "@polkadot/api";
 import { AccountId32, EventRecord } from "@polkadot/types/interfaces";
 import { Vec, u8, u32, bool } from "@polkadot/types-codec";
+import { TypeRegistry } from '@polkadot/types';
+
 export async function jumpSessions(context: DevModeContext, count: number): Promise<string | null> {
     const session = (await context.polkadotJs().query.session.currentIndex()).addn(count.valueOf()).toNumber();
 
@@ -186,6 +188,9 @@ export function fetchIssuance(events: EventRecord[] = []) {
         ({ event }: EventRecord) => event.data as unknown as { amount: u128 }
     );
 
+    if (filtered.length == 0) {
+        return  { amount: (new TypeRegistry()).createType('u128', 0) }
+    }
     return filtered[0];
 }
